servicio_auth/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic
pydantic-settings
python-dotenv
bcrypt
python-jose[cryptography]
python-multipart
pytest
httpx
email-validator